アニメーション追加

+++ Pac-ManLib/Animation.h
@@ -6,7 +6,7 @@ PTR( Animation );
 
 class Animation {
 public:
-	Animation( int anim_frame, int anim_max_time );
+	Animation( int anim_frame, double anim_max_time );
 	virtual ~Animation( );
 public:
 	void update( );
@@ -16,7 +16,7 @@ protected:
 	void setAnimationGraph( int id );
 private:
 	int _animation_fream;
-	int _animation_max_time;
-	int _animation_time;
+	double _animation_max_time;
+	double _animation_time;
 	std::vector< int > _animation_graph;
 };
